+++ src/app/modules/services/services.module.ts
@@ -1,7 +1,6 @@
 import { Module } from '@nestjs/common';
 import { MulterModule } from '@nestjs/platform-express';
 import { CqrsModule } from '@nestjs/cqrs';
-import { HttpModule } from '@nestjs/axios';
 
 import { AppConfigModule } from '@app/modules/config';
 import { SocketModule } from '@app/modules/socket';
@@ -12,7 +11,7 @@ import { CqrsHandlers } from '@domain/wasm';
 import { ServicesController } from './services.controller';
 
 @Module({
-  imports: [AppConfigModule, SocketModule, CqrsModule, HttpModule, MulterModule.register()],
+  imports: [AppConfigModule, SocketModule, CqrsModule, MulterModule.register()],
   controllers: [ServicesController],
   providers: [
     WasmService,